+++ src/codeql.test.ts
@@ -5,7 +5,7 @@ import * as toolrunner from "@actions/exec/lib/toolrunner";
 import * as io from "@actions/io";
 import * as toolcache from "@actions/tool-cache";
 import test, { ExecutionContext } from "ava";
-import del from "del";
+import * as del from "del";
 import * as yaml from "js-yaml";
 import nock from "nock";
 import * as sinon from "sinon";
@@ -18,6 +18,7 @@ import {
   AugmentationProperties,
   Config,
   defaultAugmentationProperties,
+  generateCodeScanningConfig,
 } from "./config-utils";
 import * as defaults from "./defaults.json";
 import { DocUrl } from "./doc-url";
@@ -502,8 +503,11 @@ const injectedConfigMacro = test.macro({
         ...stubConfig,
         ...configOverride,
         tempDir,
-        augmentationProperties,
       };
+      thisStubConfig.computedConfig = generateCodeScanningConfig(
+        thisStubConfig.originalUserInput,
+        augmentationProperties,
+      );
 
       await codeqlObject.databaseInitCluster(
         thisStubConfig,
@@ -523,7 +527,7 @@ const injectedConfigMacro = test.macro({
       const augmentedConfig = yaml.load(fs.readFileSync(configFile, "utf8"));
       t.deepEqual(augmentedConfig, expectedConfig);
 
-      await del(configFile, { force: true });
+      await del.deleteAsync(configFile, { force: true });
     });
   },
 
@@ -1012,7 +1016,7 @@ test("Avoids duplicating --overwrite flag if specified in CODEQL_ACTION_EXTRA_OP
   );
   t.truthy(configArg, "Should have injected a codescanning config");
   const configFile = configArg!.split("=")[1];
-  await del(configFile, { force: true });
+  await del.deleteAsync(configFile, { force: true });
 });
 
 export function stubToolRunnerConstructor(