+++ src/config-utils.test.ts
@@ -157,17 +157,17 @@ test("load empty config", async (t) => {
       }),
     );
 
-    t.deepEqual(
-      config,
-      await configUtils.getDefaultConfig(
-        createTestInitConfigInputs({
-          languagesInput: languages,
-          tempDir,
-          codeql,
-          logger,
-        }),
-      ),
+    const expectedConfig = await configUtils.initActionState(
+      createTestInitConfigInputs({
+        languagesInput: languages,
+        tempDir,
+        codeql,
+        logger,
+      }),
+      {},
     );
+
+    t.deepEqual(config, expectedConfig);
   });
 });
 
@@ -322,18 +322,21 @@ test("load non-empty input", async (t) => {
 
     fs.mkdirSync(path.join(tempDir, "foo"));
 
+    const userConfig: configUtils.UserConfig = {
+      name: "my config",
+      "disable-default-queries": true,
+      queries: [{ uses: "./foo" }],
+      "paths-ignore": ["a", "b"],
+      paths: ["c/d"],
+    };
+
     // And the config we expect it to parse to
     const expectedConfig: configUtils.Config = {
       analysisKinds: [AnalysisKind.CodeScanning],
       languages: [KnownLanguage.javascript],
       buildMode: BuildMode.None,
-      originalUserInput: {
-        name: "my config",
-        "disable-default-queries": true,
-        queries: [{ uses: "./foo" }],
-        "paths-ignore": ["a", "b"],
-        paths: ["c/d"],
-      },
+      originalUserInput: userConfig,
+      computedConfig: userConfig,
       tempDir,
       codeQLCmd: codeql.getPath(),
       gitHubVersion: githubVersion,
@@ -341,10 +344,12 @@ test("load non-empty input", async (t) => {
       debugMode: false,
       debugArtifactName: "my-artifact",
       debugDatabaseName: "my-db",
-      augmentationProperties: configUtils.defaultAugmentationProperties,
       trapCaches: {},
       trapCacheDownloadTime: 0,
       dependencyCachingEnabled: CachingKind.None,
+      extraQueryExclusions: [],
+      overlayDatabaseMode: OverlayDatabaseMode.None,
+      useOverlayDatabaseCaching: false,
     };
 
     const languagesInput = "javascript";
@@ -763,7 +768,6 @@ const calculateAugmentationMacro = test.macro({
     _title: string,
     rawPacksInput: string | undefined,
     rawQueriesInput: string | undefined,
-    rawQualityQueriesInput: string | undefined,
     languages: Language[],
     expectedAugmentationProperties: configUtils.AugmentationProperties,
   ) => {
@@ -771,7 +775,6 @@ const calculateAugmentationMacro = test.macro({
       await configUtils.calculateAugmentation(
         rawPacksInput,
         rawQueriesInput,
-        rawQualityQueriesInput,
         languages,
       );
     t.deepEqual(actualAugmentationProperties, expectedAugmentationProperties);
@@ -782,7 +785,6 @@ const calculateAugmentationMacro = test.macro({
 test(
   calculateAugmentationMacro,
   "All empty",
-  undefined,
   undefined,
   undefined,
   [KnownLanguage.javascript],
@@ -796,7 +798,6 @@ test(
   "With queries",
   undefined,
   " a, b , c, d",
-  undefined,
   [KnownLanguage.javascript],
   {
     ...configUtils.defaultAugmentationProperties,
@@ -809,7 +810,6 @@ test(
   "With queries combining",
   undefined,
   "   +   a, b , c, d ",
-  undefined,
   [KnownLanguage.javascript],
   {
     ...configUtils.defaultAugmentationProperties,
@@ -820,46 +820,8 @@ test(
 
 test(
   calculateAugmentationMacro,
-  "With quality queries",
-  undefined,
-  undefined,
-  " a, b , c, d",
-  [KnownLanguage.javascript],
-  {
-    ...configUtils.defaultAugmentationProperties,
-    qualityQueriesInput: [
-      { uses: "a" },
-      { uses: "b" },
-      { uses: "c" },
-      { uses: "d" },
-    ],
-  },
-);
-
-test(
-  calculateAugmentationMacro,
-  "With security and quality queries",
-  undefined,
-  " a, b , c, d",
-  "e, f , g,h",
-  [KnownLanguage.javascript],
-  {
-    ...configUtils.defaultAugmentationProperties,
-    queriesInput: [{ uses: "a" }, { uses: "b" }, { uses: "c" }, { uses: "d" }],
-    qualityQueriesInput: [
-      { uses: "e" },
-      { uses: "f" },
-      { uses: "g" },
-      { uses: "h" },
-    ],
-  },
-);
-
-test(
-  calculateAugmentationMacro,
   "With packs",
   "   codeql/a , codeql/b   , codeql/c  , codeql/d  ",
-  undefined,
   undefined,
   [KnownLanguage.javascript],
   {
@@ -872,7 +834,6 @@ test(
   calculateAugmentationMacro,
   "With packs combining",
   "   +   codeql/a, codeql/b, codeql/c, codeql/d",
-  undefined,
   undefined,
   [KnownLanguage.javascript],
   {
@@ -888,7 +849,6 @@ const calculateAugmentationErrorMacro = test.macro({
     _title: string,
     rawPacksInput: string | undefined,
     rawQueriesInput: string | undefined,
-    rawQualityQueriesInput: string | undefined,
     languages: Language[],
     expectedError: RegExp | string,
   ) => {
@@ -897,7 +857,6 @@ const calculateAugmentationErrorMacro = test.macro({
         configUtils.calculateAugmentation(
           rawPacksInput,
           rawQueriesInput,
-          rawQualityQueriesInput,
           languages,
         ),
       { message: expectedError },
@@ -911,7 +870,6 @@ test(
   "Plus (+) with nothing else (queries)",
   undefined,
   "   +   ",
-  undefined,
   [KnownLanguage.javascript],
   /The workflow property "queries" is invalid/,
 );
@@ -920,7 +878,6 @@ test(
   calculateAugmentationErrorMacro,
   "Plus (+) with nothing else (packs)",
   "   +   ",
-  undefined,
   undefined,
   [KnownLanguage.javascript],
   /The workflow property "packs" is invalid/,
@@ -931,7 +888,6 @@ test(
   "Packs input with multiple languages",
   "   +  a/b, c/d ",
   undefined,
-  undefined,
   [KnownLanguage.javascript, KnownLanguage.java],
   /Cannot specify a 'packs' input in a multi-language analysis/,
 );
@@ -941,7 +897,6 @@ test(
   "Packs input with no languages",
   "   +  a/b, c/d ",
   undefined,
-  undefined,
   [],
   /No languages specified/,
 );
@@ -950,7 +905,6 @@ test(
   calculateAugmentationErrorMacro,
   "Invalid packs",
   " a-pack-without-a-scope ",
-  undefined,
   undefined,
   [KnownLanguage.javascript],
   /"a-pack-without-a-scope" is not a valid pack/,