+++ src/init-action.ts
@@ -51,7 +51,9 @@ import { getRepositoryNwo } from "./repository";
 import { ToolsSource } from "./setup-codeql";
 import {
   ActionName,
-  StatusReportBase,
+  InitStatusReport,
+  InitWithConfigStatusReport,
+  createInitWithConfigStatusReport,
   createStatusReportBase,
   getActionsStatus,
   sendStatusReport,
@@ -75,52 +77,9 @@ import {
   ConfigurationError,
   wrapError,
   checkActionVersion,
-  cloneObject,
   getErrorMessage,
 } from "./util";
 import { validateWorkflow } from "./workflow";
-/** Fields of the init status report that can be sent before `config` is populated. */
-interface InitStatusReport extends StatusReportBase {
-  /** Value given by the user as the "tools" input. */
-  tools_input: string;
-  /** Version of the bundle used. */
-  tools_resolved_version: string;
-  /** Where the bundle originated from. */
-  tools_source: ToolsSource;
-  /** Comma-separated list of languages specified explicitly in the workflow file. */
-  workflow_languages: string;
-}
-
-/** Fields of the init status report that are populated using values from `config`. */
-interface InitWithConfigStatusReport extends InitStatusReport {
-  /** Comma-separated list of languages where the default queries are disabled. */
-  disable_default_queries: string;
-  /** Comma-separated list of paths, from the 'paths' config field. */
-  paths: string;
-  /** Comma-separated list of paths, from the 'paths-ignore' config field. */
-  paths_ignore: string;
-  /** Comma-separated list of queries sources, from the 'queries' config field or workflow input. */
-  queries: string;
-  /** Stringified JSON object of packs, from the 'packs' config field or workflow input. */
-  packs: string;
-  /** Comma-separated list of languages for which we are using TRAP caching. */
-  trap_cache_languages: string;
-  /** Size of TRAP caches that we downloaded, in bytes. */
-  trap_cache_download_size_bytes: number;
-  /** Time taken to download TRAP caches, in milliseconds. */
-  trap_cache_download_duration_ms: number;
-  /** Size of the overlay-base database that we downloaded, in bytes. */
-  overlay_base_database_download_size_bytes?: number;
-  /** Time taken to download the overlay-base database, in milliseconds. */
-  overlay_base_database_download_duration_ms?: number;
-  /** Stringified JSON array of registry configuration objects, from the 'registries' config field
-  or workflow input. **/
-  registries: string;
-  /** Stringified JSON object representing a query-filters, from the 'query-filters' config field. **/
-  query_filters: string;
-  /** Path to the specified code scanning config file, from the 'config-file' config field. */
-  config_file: string;
-}
 
 /** Fields of the init status report populated when the tools source is `download`. */
 interface InitToolsDownloadFields {
@@ -180,83 +139,17 @@ async function sendCompletedStatusReport(
   }
 
   if (config !== undefined) {
-    const languages = config.languages.join(",");
-    const paths = (config.originalUserInput.paths || []).join(",");
-    const pathsIgnore = (config.originalUserInput["paths-ignore"] || []).join(
-      ",",
-    );
-    const disableDefaultQueries = config.originalUserInput[
-      "disable-default-queries"
-    ]
-      ? languages
-      : "";
-
-    const queries: string[] = [];
-    let queriesInput = getOptionalInput("queries")?.trim();
-    if (queriesInput === undefined || queriesInput.startsWith("+")) {
-      queries.push(
-        ...(config.originalUserInput.queries || []).map((q) => q.uses),
-      );
-    }
-    if (queriesInput !== undefined) {
-      queriesInput = queriesInput.startsWith("+")
-        ? queriesInput.slice(1)
-        : queriesInput;
-      queries.push(...queriesInput.split(","));
-    }
-
-    let packs: Record<string, string[]> = {};
-    if (
-      (config.augmentationProperties.packsInputCombines ||
-        !config.augmentationProperties.packsInput) &&
-      config.originalUserInput.packs
-    ) {
-      // Make a copy, because we might modify `packs`.
-      const copyPacksFromOriginalUserInput = cloneObject(
-        config.originalUserInput.packs,
-      );
-      // If it is an array, then assume there is only a single language being analyzed.
-      if (Array.isArray(copyPacksFromOriginalUserInput)) {
-        packs[config.languages[0]] = copyPacksFromOriginalUserInput;
-      } else {
-        packs = copyPacksFromOriginalUserInput;
-      }
-    }
-
-    if (config.augmentationProperties.packsInput) {
-      packs[config.languages[0]] ??= [];
-      packs[config.languages[0]].push(
-        ...config.augmentationProperties.packsInput,
-      );
-    }
-
     // Append fields that are dependent on `config`
-    const initWithConfigStatusReport: InitWithConfigStatusReport = {
-      ...initStatusReport,
-      config_file: configFile ?? "",
-      disable_default_queries: disableDefaultQueries,
-      paths,
-      paths_ignore: pathsIgnore,
-      queries: queries.join(","),
-      packs: JSON.stringify(packs),
-      trap_cache_languages: Object.keys(config.trapCaches).join(","),
-      trap_cache_download_size_bytes: Math.round(
-        await getTotalCacheSize(Object.values(config.trapCaches), logger),
-      ),
-      trap_cache_download_duration_ms: Math.round(config.trapCacheDownloadTime),
-      overlay_base_database_download_size_bytes:
-        overlayBaseDatabaseStats?.databaseSizeBytes,
-      overlay_base_database_download_duration_ms:
-        overlayBaseDatabaseStats?.databaseDownloadDurationMs,
-      query_filters: JSON.stringify(
-        config.originalUserInput["query-filters"] ?? [],
-      ),
-      registries: JSON.stringify(
-        configUtils.parseRegistriesWithoutCredentials(
-          getOptionalInput("registries"),
-        ) ?? [],
-      ),
-    };
+    const initWithConfigStatusReport: InitWithConfigStatusReport =
+      await createInitWithConfigStatusReport(
+        config,
+        initStatusReport,
+        configFile,
+        Math.round(
+          await getTotalCacheSize(Object.values(config.trapCaches), logger),
+        ),
+        overlayBaseDatabaseStats,
+      );
     await sendStatusReport({
       ...initWithConfigStatusReport,
       ...initToolsDownloadFields,
@@ -384,11 +277,21 @@ async function run() {
       }
     }
 
+    // Warn that `quality-queries` is deprecated if there is an argument for it.
+    const qualityQueriesInput = getOptionalInput("quality-queries");
+
+    if (qualityQueriesInput !== undefined) {
+      logger.warning(
+        "The `quality-queries` input is deprecated and will be removed in a future version of the CodeQL Action. " +
+          "Use the `analysis-kinds` input to configure different analysis kinds instead.",
+      );
+    }
+
     config = await initConfig({
       analysisKindsInput: getRequiredInput("analysis-kinds"),
       languagesInput: getOptionalInput("languages"),
       queriesInput: getOptionalInput("queries"),
-      qualityQueriesInput: getOptionalInput("quality-queries"),
+      qualityQueriesInput,
       packsInput: getOptionalInput("packs"),
       buildModeInput: getOptionalInput("build-mode"),
       configFile,
@@ -439,9 +342,8 @@ async function run() {
   let overlayBaseDatabaseStats: OverlayBaseDatabaseDownloadStats | undefined;
   try {
     if (
-      config.augmentationProperties.overlayDatabaseMode ===
-        OverlayDatabaseMode.Overlay &&
-      config.augmentationProperties.useOverlayDatabaseCaching
+      config.overlayDatabaseMode === OverlayDatabaseMode.Overlay &&
+      config.useOverlayDatabaseCaching
     ) {
       // OverlayDatabaseMode.Overlay comes in two flavors: with database
       // caching, or without. The flavor with database caching is intended to be
@@ -460,8 +362,7 @@ async function run() {
         logger,
       );
       if (!overlayBaseDatabaseStats) {
-        config.augmentationProperties.overlayDatabaseMode =
-          OverlayDatabaseMode.None;
+        config.overlayDatabaseMode = OverlayDatabaseMode.None;
         logger.info(
           "No overlay-base database found in cache, " +
             `reverting overlay database mode to ${OverlayDatabaseMode.None}.`,
@@ -469,10 +370,7 @@ async function run() {
       }
     }
 
-    if (
-      config.augmentationProperties.overlayDatabaseMode !==
-      OverlayDatabaseMode.Overlay
-    ) {
+    if (config.overlayDatabaseMode !== OverlayDatabaseMode.Overlay) {
       cleanupDatabaseClusterDirectory(config, logger);
     }
 
@@ -729,15 +627,13 @@ async function run() {
     // revert to `OverlayDatabaseMode.None`, re-initialize the database cluster
     // with the new overlay database mode.
     if (
-      config.augmentationProperties.overlayDatabaseMode !==
-        OverlayDatabaseMode.None &&
+      config.overlayDatabaseMode !== OverlayDatabaseMode.None &&
       !(await checkPacksForOverlayCompatibility(codeql, config, logger))
     ) {
       logger.info(
         "Reverting overlay database mode to None due to incompatible packs.",
       );
-      config.augmentationProperties.overlayDatabaseMode =
-        OverlayDatabaseMode.None;
+      config.overlayDatabaseMode = OverlayDatabaseMode.None;
       cleanupDatabaseClusterDirectory(config, logger, {
         disableExistingDirectoryWarning: true,
       });